example/requirements/development.txt
-r base.txt
django-debug-toolbar==0.9.4
ipdb==0.7
django-pdb==0.3.2
coverage==3.5.3
django-coverage==1.2.2
docutils==0.9.1
selenium==2.25.0
